challenge_semana4_Castro_Wachs_Trinidad/requirements.txt
cohere
python-dotenv
langchain
fastapi
pydantic
uuid
scikit-learn
uvicorn   
